Update proposal api

+++ backend/models/proposal.go
@@ -3,10 +3,10 @@ package models
 import "time"
 
 type Proposal struct {
-	ID             int       `json:"id"`
+	ID             int64     `json:"id"`
 	FileURL        string    `json:"file_url"`
 	CreatedAt      time.Time `json:"created_at"`
 	UpdatedAt      time.Time `json:"updated_at"`
 	StatusProposal string    `json:"status_proposal"`
-	UserID         int       `json:"user_id"`
+	UserID         int64     `json:"user_id"`
 }
